# docs/swagger.yaml
definitions:
  main.Users:
    properties:
      age:
        description: Age
        example: 10
        type: integer
      id:
        description: UserId
        example: 1
        type: integer
      name:
        description: Name
        example: John
        type: string
    type: object
  swagger_search.APIError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
  swagger_search.Search:
    properties:
      age:
        description: Age
        example: 10
        type: integer
      id:
        description: UserId
        example: 1
        type: integer
      name:
        description: Name
        example: John
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: test swagger api
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: test swagger api
  /es/search:
    post:
      consumes:
      - application/json
      description: search engine api
      parameters:
      - description: Search Info Body
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/swagger_search.Search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger_search.Search'
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/swagger_search.APIError'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/swagger_search.APIError'
      summary: search engine api
      tags:
      - Search
  /hello/{name}:
    get:
      consumes:
      - application/json
      description: test swagger api
      parameters:
      - description: Users name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Users'
        "400":
          description: Bad Request
      summary: test swagger api
swagger: "2.0"